requirements.txt
scipy==1.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
et-xmlfile==1.1.0
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4  
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
openai==0.19.0
openpyxl==3.0.10
pandas==1.4.2
pandas-stubs==1.2.0.58
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
